Added file compatibility tests for config parsing
