+++ packages/server/src/config/readConfig.tsx
@@ -1,0 +1,28 @@
+import { Config } from 'convict';
+import { pathExistsSync } from 'fs-extra';
+import { join } from 'path';
+import { schema } from './schema';
+
+export function readConfig<T>(): Config<T> {
+  const cwd = process.cwd();
+  let parsedConfig: Config<T> | undefined;
+  const extensions = ['json', 'yaml', 'yml', 'json5'];
+  for (const ext of extensions) {
+    try {
+      const path = join(cwd, `openapi-platform.config.${ext}`);
+      if (pathExistsSync(path)) {
+        parsedConfig = schema.loadFile(path);
+        break;
+      }
+    } catch (err) {
+      parsedConfig = undefined;
+    }
+  }
+  if (!parsedConfig) {
+    throw new Error(
+      `Failed to find an openapi-platform.config.${extensions.join('/')} file`,
+    );
+  }
+  schema.validate({ allowed: 'strict' });
+  return parsedConfig;
+}
