+++ packages/server/test/system/config/config.yml
@@ -1,0 +1,6 @@
+database:
+  name: test_name
+  host: test_host
+  port: 1337
+  username: test_username
+  password: test_password
